fix(gui): 修复类型转换和面包屑导航问题

+++ file/directory.h
@@ -73,7 +73,7 @@ class path_manager
     }
     bool navigate_to_breadcrumb(int index)
     {
-        if (index < 0 || index >= paths_.size())
+        if (index < 0 || index >= static_cast<int>(paths_.size()))
         {
             return false;
         }
